requirements.txt
flask>=2.3.0
flask-dotenv>=0.1.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0
requests>=2.31.0
psycopg2-binary>=2.9.9
